Add files via upload
The Pig game is a simple dice game in which players take turns rolling a single die to accumulate points. The objective is to be the first player to reach a certain score (50 points). However, players must be cautious as they roll the dice because if they roll a 1, they lose all the points they've accumulated in that turn, and their turn ends.